refactor index for shared api
