--- src/shared/api/index.ts
@@ -1,10 +1,0 @@
-export * from './auth/authApi';
-export * from './common/model/api.types';
-export * from './mainPhotoProfile/mainPhotoProfileAPI';
-export * from './personalInformationUser/personalInformationUserAPI';
-export * from './posts/postsApi';
-export * from './posts/postsTypes';
-export * from './profile/profileApi';
-export * from './profile/profileTypes';
-export * from './public/publicApi';
-export * from './public/publicTypes';
