+++ src/shared/api/lib/transformErrorResponse.ts
@@ -1,4 +1,4 @@
-import { BaseResponseType } from '@/shared/api';
+import { BaseResponseType } from '@/shared/api/common/model/api.types';
 
 export const transformErrorResponse = (res: { data: BaseResponseType; status: number }) => {
   return res.data;
