+++ src/widget/myProfile/MyProfile.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useGetPersonalInformationUserQuery } from '@/shared/api/';
+import { useGetPersonalInformationUserQuery } from '@/shared/api/personalInformationUser/personalInformationUserAPI';
 import { PhotoProfile, Post, Typography, Wrapper } from '@/shared/ui';
 import Link from 'next/link';
