+++ src/widget/postInteractionBar/PostInteractionBar.tsx
@@ -1,5 +1,8 @@
 import { ReSendPost, ToggleFavorites, ToggleLike } from '@/features';
-import { GetPostByIdResponse, GetPostLikesResponse, useMeQuery, useUpdateLikePostMutation } from '@/shared/api';
+import { useMeQuery } from '@/shared/api/auth/authApi';
+import { useUpdateLikePostMutation } from '@/shared/api/posts/postsApi';
+import { GetPostLikesResponse } from '@/shared/api/posts/postsTypes';
+import { GetPostByIdResponse } from '@/shared/api/public/publicTypes';
 import { LikeStatus } from '@/shared/lib';
 
 import s from './PostInteractionBar.module.scss';
